test(ci): split ci-sync.e2e.ts into five files, and address the second review round
The e2e splitter assigns whole files to nodes (scripts/split-e2e-tests.js), so
a single large file sets the floor for the job. ci-sync.e2e.ts had grown to
1722 lines and 21 top-level blocks; it is now five files along its existing
seams - lane-cycle, main, branch-ownership, convergence, scopes - with all 57
tests unchanged. armPrePushHook moves to the shared support module. The
manifest entry is redistributed across the five by measured per-file ratios,
keeping the CI-scale total.

Review round two:
- dropped the eslint-disable-next-line for no-await-in-loop: the rule is not
  configured in .oxlintrc.json, so the directive suppressed nothing.
- importHeadsOf now carries the import error instead of returning a bare
  boolean, so a failed head fetch reports its cause - the state where every
  entry is left untouched is exactly when that cause is needed.
