+++ scripts/e2e-test-timings.json
@@ -50,8 +50,12 @@
   "e2e/harmony/checkout-harmony.e2e.ts": 249,
   "e2e/harmony/ci-bitmap-auto-sync.e2e.ts": 587,
   "e2e/harmony/ci-commands.e2e.ts": 590,
+  "e2e/harmony/ci-sync-branch-ownership.e2e.ts": 54,
+  "e2e/harmony/ci-sync-convergence.e2e.ts": 90,
+  "e2e/harmony/ci-sync-lane-cycle.e2e.ts": 56,
+  "e2e/harmony/ci-sync-main.e2e.ts": 53,
+  "e2e/harmony/ci-sync-scopes.e2e.ts": 47,
   "e2e/harmony/ci-sync-state.e2e.ts": 96,
-  "e2e/harmony/ci-sync.e2e.ts": 298,
   "e2e/harmony/compile.e2e.ts": 15,
   "e2e/harmony/component-config.e2e.ts": 23,
   "e2e/harmony/corrupted-objects.e2e.ts": 15,
